requirements.txt
requests
flask
python-dotenv
pandas
altair